Resolve clippy lints across crate
Clean up clippy warnings and deny-by-default errors on the default,
visual-backend, and pgm-backend feature sets:

- Replace tautological `bool || !bool` test asserts with no-panic calls
  (overly_complex_bool_expr); drop dead key_states vec.
- Derive Default instead of manual impls; add Default for MiniFbBackend
  and GraphicsMode (derivable_impls, new_without_default,
  should_implement_trait).
- Make thread_local initializers const; use range contains and
  is_some_and; vec![] over Vec::new()+push; struct-init WindowOptions;
  or_default over or_insert_with(HashMap::new).
- Replace `if let Err(_) = x {}` with `let _ = x;` / .is_err();
  collapse if-let into let-chains.
- Drop redundant bindings/field names, unnecessary parens, unused
  import, .clone() on Copy; println!() over println!("").
- #[allow(too_many_arguments)] on coherent draw-primitive signatures;
  BatchOp type alias for complex type.

winit-backend remains out of sync with the Backend trait and is not
addressed here.

+++ tests/integration/test_visual_output.rs
@@ -318,7 +318,7 @@ fn test_visual_performance_basic() {
 
     // Perform basic graphics operations
     for i in 0..1000 {
-        putpixel((i % 100) as i32, ((i / 100) % 100) as i32, Color::WHITE);
+        putpixel(i % 100, (i / 100) % 100, Color::WHITE);
     }
 
     let duration = start.elapsed();
